complete digital thermometer firmware

+++ src/filter.c
@@ -2,9 +2,13 @@
 
 #include "filter.h"
 
+
 static uint16_t sample_sum = 0;
+
 static uint8_t sample_count = 0;
+
 static uint8_t filter_ready = 0;
+
 static uint8_t filtered_value = 0;
 
 
@@ -12,7 +16,9 @@ void Filter_Reset(void)
 {
     sample_sum = 0;
     sample_count = 0;
+
     filter_ready = 0;
+
     filtered_value = 0;
 }
 
@@ -23,11 +29,20 @@ void Filter_AddSample(uint8_t sample)
 
     sample_count++;
 
+
     if (sample_count >= FILTER_SAMPLES)
     {
-        filtered_value = sample_sum >> 7;
+        /*
+         * Division by 128.
+         *
+         * 128 = 2^7
+         */
+        filtered_value =
+            (uint8_t)(sample_sum >> 7);
+
 
         sample_sum = 0;
+
         sample_count = 0;
 
         filter_ready = 1;
